结构优化
1. 删掉匿名方法,使用专门的命令解析
2. 结构优化,尽可能保证单一职责
