+++ .obsidian/workspace.json
@@ -168,11 +168,20 @@
       "command-palette:開啟命令面板": false
     }
   },
-  "active": "67e95133d30fca46",
+  "active": "b13273279c5196ad",
   "lastOpenFiles": [
+    "WeChatBot.Console/si0wtjje.2sa~",
+    "WeChatBot.Console/Commands/us5sgl20.pxt~",
+    "WeChatBot.Console/rg3ieste.oq5~",
+    "WeChatBot.Console/cbxzw5au.lvh~",
+    "WeChatBot.Console/44kslsdo.fmv~",
+    "WeChatBot.Console/npax2s41.s13~",
+    "WeChatBot.Console/sqjg0d40.xbt~",
+    "WeChatBot.Console/v2zh1shg.ynq~",
+    "WeChatBot.Console/mxzcfcza.m5t~",
+    "WeChatBot.Console/12f2tm1v.jp2~",
+    "WeChatBot.Console/2c121p1a.mcu~",
     "WeChatBot.Example/bot_1.png",
-    "WeChatBot.Example/微信圖片_20250802172909.png",
-    "WeChatBot.Example",
-    "新建文件夹"
+    "WeChatBot.Example/微信圖片_20250802172909.png"
   ]
 }
